fix geodata typing and cyclic dependency
