+++ packages/common/package.json
@@ -13,7 +13,6 @@
     "test:watch": "vitest"
   },
   "dependencies": {
-    "clickhouse": "workspace:*",
     "hono": "^4.6.13",
     "nanoid": "^5.0.4",
     "pino": "^9.6.0",
