+++ packages/queues/package.json
@@ -12,7 +12,6 @@
   "dependencies": {
     "@vemetric/logger": "workspace:*",
     "bullmq": "^5.73.1",
-    "clickhouse": "workspace:*",
     "database": "workspace:*",
     "ioredis": "^5.10.1",
     "zod": "3.24.2"
